requirements.txt
Flask==2.2.3
Flask-PyMongo==2.3.0
connexion[swagger-ui]==2.14.1
pymongo==4.3.3
marshmallow==3.19.0
python-dotenv==1.0.0
Werkzeug==2.2.3
Flask-JWT-Extended==4.4.4